+++ angular/src/lib/call-control/index.ts
@@ -1,2 +1,0 @@
-export * from './call-control.module';
-export * from './call-control.component';