+++ angular/src/lib/checkbox/examples/index.ts
@@ -1,7 +1,0 @@
-export * from './default.component';
-export * from './disabled.component';
-export * from './indeterminate.component';
-export * from './nested.component';
-export * from './kitchen-sink.component';
-export * from './single.component';
-export * from './forms.component';