+++ angular/src/lib/input/index.ts
@@ -1,7 +1,0 @@
-export * from '../input/input.directive';
-export * from '../input/input.service';
-export * from '../input-container/input-container.component';
-export * from '../input-helper/input-helper.component';
-export * from '../input-message/input-message.component';
-export * from '../input-section/input-section.component';
-export * from './input.module';