+++ angular/src/lib/list/index.ts
@@ -1,2 +1,0 @@
-export * from './list.module';
-export * from './list.component';