+++ angular/src/lib/select/index.ts
@@ -1,2 +1,0 @@
-export * from './select.module';
-export * from './select.component';