+++ angular/src/lib/tabs/examples/index.ts
@@ -1,3 +1,0 @@
-export * from './default.component';
-export * from './disable.component';
-export * from './pills.component';