+++ angularjs/src/lib/services/multimap/index.ts
@@ -1,7 +1,0 @@
-import * as angular from 'angular';
-import { MdMultiMap } from './multimap.service';
-
-export default angular
-  .module('momentum.ui.multimap', [])
-  .factory('$$multiMap', MdMultiMap)
-  .name;