+++ angularjs/src/lib/services/position/index.ts
@@ -1,7 +1,0 @@
-import * as angular from 'angular';
-import { $position } from './position.service';
-
-export default angular
-  .module('momentum.ui.position', [])
-  .factory('$position', $position)
-  .name;