+++ vue/src/lib/avatar/examples/image.vue
@@ -1,9 +1,0 @@
-<template>
-  <md-avatar title="Tom Smith" src="http://react.collab-ui.com/barbara.png" />
-</template>
-
-<script>
-export default {
-  name: 'ExampleAvatarImage',
-};
-</script>