+++ vue/src/lib/input/index.js
@@ -1,7 +1,0 @@
-import Input from './index.vue';
-
-Input.install = function(Vue) {
-  Vue.component(Input.name, Input);
-};
-
-export default Input;