+++ vue/src/lib/sidebar/sidebar-header/index.js
@@ -1,7 +1,0 @@
-import SidebarHeader from './index.vue';
-
-SidebarHeader.install = function(Vue) {
-  Vue.component(SidebarHeader.name, SidebarHeader);
-};
-
-export default SidebarHeader;