+++ vue/src/lib/tabs/index.js
@@ -1,7 +1,0 @@
-import Tabs from './index.vue';
-
-Tabs.install = function(Vue) {
-  Vue.component(Tabs.name, Tabs);
-};
-
-export default Tabs;